+++ src/components/Search/index.tsx
@@ -28,7 +28,7 @@ import { Suggestion, SuggestionType, DateValue, Bbox } from 'types/search';
 import { Search, SearchStats } from 'types/gallery';
 import { FILE_TYPE } from 'constants/file';
 import { GalleryContext } from 'pages/gallery';
-import { SelectStyles } from './SelectStyle';
+import { SelectStyles } from './styles';
 
 const Wrapper = styled.div<{ isDisabled: boolean; isOpen: boolean }>`
     position: fixed;
