+++ src/components/Search/styles.ts
